refactored structure, removed makefile, added .editorconfig
