--- Makefile
@@ -1,5 +1,0 @@
-
-test:
-	@node test/test.migrate.js
-
-.PHONY: test
