+++ lib/set.js
@@ -1,3 +1,4 @@
+'use strict'
 
 /*!
  * migrate - Set
